+++ services/web/server/src/simcore_service_webserver/exporter/_formatter/xlsx/utils.py
@@ -22,7 +22,7 @@ def get_max_array_length(*arrays_of_elements) -> int:
     return max(len(x) for x in arrays_of_elements)
 
 
-def column_iter(start_from: int, elements: int) -> Iterable[str]:
+def column_generator(start_from: int, elements: int) -> Iterable[str]:
     """maps columns index to letters"""
     for column_index in range(start_from, start_from + elements):
         yield get_column_letter(column_index)
