+++ services/web/server/src/simcore_service_webserver/exporter/request_handlers.py
@@ -12,8 +12,8 @@ from ..projects.project_lock import lock_project
 from ..projects.projects_api import retrieve_and_notify_project_locked_state
 from ..security.decorators import permission_required
 from ..users.api import get_user_name
+from ._formatter.archive import get_sds_archive_path
 from .exceptions import SDSException
-from .formatter.archive import get_sds_archive_path
 from .utils import CleanupFileResponse
 
 _logger = logging.getLogger(__name__)
